# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Linguastar API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /purchases:
    get:
      operationId: listPurchases
      tags: [purchases]
      summary: List current user purchases
      responses:
        "200":
          description: User purchases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Purchase"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createPurchase
      tags: [purchases]
      summary: Purchase a book
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePurchaseBody"
      responses:
        "201":
          description: Purchase recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Purchase"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /reading-progress:
    get:
      operationId: listReadingProgress
      tags: [reading-progress]
      summary: Get all reading progress for current user
      responses:
        "200":
          description: Reading progress list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReadingProgress"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /reading-progress/{bookId}:
    get:
      operationId: getReadingProgress
      tags: [reading-progress]
      summary: Get reading progress for a specific book
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reading progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadingProgress"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateReadingProgress
      tags: [reading-progress]
      summary: Update reading progress
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateReadingProgressBody"
      responses:
        "200":
          description: Progress updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadingProgress"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/purchases:
    get:
      operationId: adminListPurchases
      tags: [admin]
      summary: List all purchases (admin only)
      responses:
        "200":
          description: All purchases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Purchase"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Book:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        author:
          type: string
        description:
          type: string
        coverUrl:
          type: ["string", "null"]
        price:
          type: number
        category:
          type: string
        language:
          type: string
        level:
          type: string
        isFeatured:
          type: boolean
        isBestseller:
          type: boolean
        totalPages:
          type: integer
        rating:
          type: number
        reviewCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - title
        - author
        - description
        - price
        - category
        - language
        - level
        - isFeatured
        - isBestseller
        - totalPages
        - rating
        - reviewCount
        - createdAt
        - updatedAt
...
    Purchase:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        bookId:
          type: integer
        amount:
          type: number
        status:
          type: string
        createdAt:
          type: string
        book:
          $ref: "#/components/schemas/Book"
      required:
        - id
        - userId
        - bookId
        - amount
        - status
        - createdAt
        - book
    CreatePurchaseBody:
      type: object
      properties:
        bookId:
          type: integer
        amount:
          type: number
      required:
        - bookId
        - amount
    ReadingProgress:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        bookId:
          type: integer
        currentPage:
          type: integer
        totalPages:
          type: integer
        percentComplete:
          type: number
        lastReadAt:
          type: string
        book:
          $ref: "#/components/schemas/Book"
      required:
        - id
        - userId
        - bookId
        - currentPage
        - totalPages
        - percentComplete
        - lastReadAt
        - book
    UpdateReadingProgressBody:
      type: object
      properties:
        currentPage:
          type: integer
        totalPages:
          type: integer
      required:
        - currentPage
...